chore: moved AddDescriptionErrorLogs test cases to seperate file
